# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "myArea REST API",
    "description": ""
  },
  "host": "api.pezentics.com",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/api/user/addRelation": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "objItem": {
                  "example": "any"
                },
                "Email": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                },
                "Level": {
                  "example": "any"
                },
                "RelationId": {
                  "example": "any"
                },
                "RefId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/getAllRelationByCriteria": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "LastName": {
                  "example": "any"
                },
                "FirstName": {
                  "example": "any"
                },
                "NickName": {
                  "example": "any"
                },
                "FamilyName": {
                  "example": "any"
                },
                "Region": {
                  "example": "any"
                },
                "HomeTown": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/getRelation/{relationId}/{relationType}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "relationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "relationType",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/getAllRelation/{userId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/getAllRelationByName": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/user/deleteRelation/{relationId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "relationId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/addChildOrSibling": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "objItem": {
                  "example": "any"
                },
                "Email": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                },
                "RelationType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateChildOrSibling": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "child": {
                  "example": "any"
                },
                "Email": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                },
                "RelationType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/getChildOrSibling/{relationId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "relationId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/getAllChildorSibling/{relationType}/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "relationType",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/deleteChildorSibling/{relationId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "relationId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/addSchoolplacework": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "objItem": {
                  "example": "any"
                },
                "Email": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                },
                "RelationType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateSchoolPlaceWork": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "child": {
                  "example": "any"
                },
                "Email": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                },
                "RelationType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/getSchoolPlaceWork/{Id}/{relationType}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "relationType",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/getAllSchoolPlaceWork/{userId}/{relationType}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "relationType",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/deleteSchoolPlaceWork/{relationType}/{Id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "relationType",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/getNeighbourhoodByUserId/{userId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllUser/{userId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllUser": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findUser/{userId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllUserBySearch/{name}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllUsersByDate/{startDate}/{toDate}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateUser/{userId}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "example": "any"
                },
                "RelationType": {
                  "example": "any"
                },
                "RelationCategory": {
                  "example": "any"
                },
                "FirstName": {
                  "example": "any"
                },
                "LastName": {
                  "example": "any"
                },
                "MiddleName": {
                  "example": "any"
                },
                "NickName": {
                  "example": "any"
                },
                "Level": {
                  "example": "any"
                },
                "Email": {
                  "example": "any"
                },
                "Age": {
                  "example": "any"
                },
                "Sex": {
                  "example": "any"
                },
                "DOB": {
                  "example": "any"
                },
                "MaritalStatus": {
                  "example": "any"
                },
                "EmploymentStatus": {
                  "example": "any"
                },
                "Tribe": {
                  "example": "any"
                },
                "FamilyName": {
                  "example": "any"
                },
                "Language": {
                  "example": "any"
                },
                "Kindred": {
                  "example": "any"
                },
                "Clan": {
                  "example": "any"
                },
                "Mobile": {
                  "example": "any"
                },
                "Address": {
                  "example": "any"
                },
                "City": {
                  "example": "any"
                },
                "HomeTown": {
                  "example": "any"
                },
                "LGA": {
                  "example": "any"
                },
                "State": {
                  "example": "any"
                },
                "Country": {
                  "example": "any"
                },
                "ProfilePicture": {
                  "example": "any"
                },
                "CoverPicture": {
                  "example": "any"
                },
                "Desc": {
                  "example": "any"
                },
                "Occupation": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateFile": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "example": "any"
                },
                "FileType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/uploadCompanyDoc": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "RefId": {
                  "example": "any"
                },
                "DocTitle": {
                  "example": "any"
                },
                "DocType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/user/delete": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "token",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/deleteAll": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findRoles": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findUserRoles": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateRole/{roleId}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateUserRole/{userId}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/deleteRole/{roleId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "token",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}